Create README.md
This is a Little Project for School.

Everybody can use the code, if you not move the Copyrights. 

The License ist the GNU Public License http://opensource.org/licenses/gpl-license.php


If somebody find an error or an Bug, please let it know me. 
I would be happy about suggestions and code changes.